model_lib/data_collection/requirements.txt
python-youtube
pytube3
moviepy
youtube-transcript-api